+++ src/main/java/com/gmr/porfolio/services/UserRolService.java
@@ -19,14 +19,14 @@ public class UserRolService {
     @Autowired
     Userdao userdao;
 
-    public  ArrayList<String> getUserRoles(Long user_id) throws NoSuchAlgorithmException, InvalidKeySpecException {
+    public  ArrayList<String> getUserRoles(int user_id) throws NoSuchAlgorithmException, InvalidKeySpecException {
 
         ArrayList<String> roles = new ArrayList<>(userRoldao.getRoles(user_id));
         return roles;
 
     }
 
-    public void setUserRoles(String userName, Long userID){
+    public void setUserRoles(String userName, int userID){
         UserRol uRol = new UserRol();
         uRol.setUserID(userID);
         uRol.setRol("common");
@@ -34,7 +34,7 @@ public class UserRolService {
 
     }
 
-    public void deleteUserRoles(Long idUser, int lenRoles) throws NoSuchAlgorithmException, InvalidKeySpecException {
+    public void deleteUserRoles(int idUser, int lenRoles) throws NoSuchAlgorithmException, InvalidKeySpecException {
         for (int i=0; i< lenRoles; i++){
             userRoldao.deleteUserRol(idUser);
         }
